Baker inn tekst fra vedlegg varsel om tilbakekreving inn i brevet. Justerer navn på saksbehandlervalg og oppdaterer til riktig tekst fra lovdata.
