+++ pensjon-brevbaker-api-model/src/main/kotlin/no/nav/pensjon/brev/api/model/maler/redigerbar/VarselTilbakekrevingAvFeilutbetaltBeloepDto.kt
@@ -3,6 +3,7 @@ package no.nav.pensjon.brev.api.model.maler.redigerbar
 import no.nav.pensjon.brev.api.model.Sakstype
 import no.nav.pensjon.brev.api.model.maler.BrevbakerBrevdata
 import no.nav.pensjon.brev.api.model.maler.RedigerbarBrevdata
+import no.nav.pensjon.brevbaker.api.model.DisplayText
 
 @Suppress("unused")
 data class VarselTilbakekrevingAvFeilutbetaltBeloepDto(
@@ -12,7 +13,8 @@ data class VarselTilbakekrevingAvFeilutbetaltBeloepDto(
 
     data class SaksbehandlerValg(
         val hvisAktueltAaIleggeRentetillegg: Boolean,
-        val hvisGjenlevendepensjonOgForsoergningstillegg: Boolean
+        @DisplayText("Hvis avregning etter § 22-16")
+        val hvisAvregningEtterParagraf2216: Boolean
     ) : BrevbakerBrevdata
 
     data class PesysData(val sakstype: Sakstype) : BrevbakerBrevdata
